feat: Add SessionsLibrary
