+++ Sessions.csproj
@@ -17,7 +17,14 @@
 
     <ItemGroup>
         <None Update="lang\**\*.*" CopyToOutputDirectory="PreserveNewest" />
-        <Compile Remove="1.1\**" />
+        <Compile Remove = ".\SessionsLibrary\**" />
+        <ProjectReference Include=".\SessionsLibrary\SessionsLibrary.csproj" />
     </ItemGroup>
+
+    <Target Name="PostBuild" AfterTargets="PostBuildEvent">
+        <Exec Command="del $(OutDir)SessionsLibrary.dll" />
+        <Exec Command="del $(OutDir)SessionsLibrary.pdb" />
+        <Exec Command="del $(OutDir)SessionsLibrary.deps.json" />
+    </Target>
 
 </Project>
